Feat: Add global DarkVeil background and improve TextRoll animation
- Moved DarkVeil to App.jsx for global background
- Updated TextRoll component to handle uppercase and improve animation
- Added Inter ExtraBold font and configured in index.css
- Cleaned up individual page components

+++ frontend/src/components/DarkVeil.jsx
@@ -7,9 +7,10 @@ attribute vec2 position;
 void main(){gl_Position=vec4(position,0.0,1.0);}
 `;
 
+/* The fragment shader is quite complex, so we define it here carefully */
 const fragment = `
 #ifdef GL_ES
-precision lowp float;
+precision highp float;
 #endif
 uniform vec2 uResolution;
 uniform float uTime;
